# project-controller/src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description: new-project管理后台API
  version: 1.0.0
  title: 新项目管理后台
basePath: /
schemes:
  - http

#认证
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "token"
    in: "header"

#定义接口数据
paths:
...
  /sys/dept/check:
    get:
      tags:
        - 部门管理
      summary: 校验部门名称是否已存在
      produces:
        - application/json
      parameters:
        - name: deptName
          description: 部门名称
          in: query
          type: string
          required: true
      responses:
        '200':
          description: 返回联系方式是否存在
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              schedule:
                description: 返回部门名称是否存在
  /sys/dept/list:
    get:
      tags:
        - 部门管理
      summary: 查询系统部门列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序字段
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: deptId
          description: 部门ID
          in: query
          type: integer
        - name: deptName
          description: 部门名称
          in: query
          type: string
        - name: status
          description: 状态 0.禁用 1.正常
          in: query
          type: integer
      responses:
        '200':
          description: 返回部门列表
          schema:
            $ref: '#/definitions/SysDeptEntities'
  /sys/dept/save:
    post:
      tags:
        - 部门管理
      summary: 新增部门
      produces:
        - application/json
      parameters:
        - name: body
          description: 部门对象
          in: body
          schema:
            $ref: '#/definitions/SysDeptSaveVo'
          required: true
      responses:
        '200':
          description: 返回信息
          schema:
            $ref: '#/definitions/R'
  /sys/dept/info/{deptId}:
    get:
      tags:
        - 部门管理
      summary: 获取部门信息
      produces:
        - application/json
      parameters:
        - name: deptId
          description: 部门ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回部门信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              user:
                $ref: '#/definitions/SysDeptInfoVo'
  /sys/dept/update:
    post:
      tags:
        - 部门管理
      summary: 修改部门
      produces:
        - application/json
      parameters:
        - name: body
          description: 部门对象
          in: body
          schema:
            $ref: '#/definitions/SysDeptUpdateVo'
          required: true
      responses:
        '200':
          description: 返回信息
          schema:
            $ref: '#/definitions/R'
  /sys/dept/normal/{deptId}:
    get:
      tags:
        - 部门管理
      summary: 修改部门状态为启用
      produces:
        - application/json
      parameters:
        - name: deptId
          description: 部门ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回信息
          schema:
            $ref: '#/definitions/R'
  /sys/dept/disable/{deptId}:
    get:
      tags:
        - 部门管理
      summary: 修改部门状态为禁用
      produces:
        - application/json
      parameters:
        - name: deptId
          description: 部门ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回信息
          schema:
            $ref: '#/definitions/R'
  /sys/dept/delete/{deptId}:
    get:
      tags:
        - 部门管理
      summary: 删除部门
      produces:
        - application/json
      parameters:
        - name: deptId
          description: 部门ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回信息
          schema:
            $ref: '#/definitions/R'
...
  /cou/wares/list:
    get:
      tags:
        - 品牌管理-商品模块
      summary: 分页查询商品列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序字段
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: couBrandId
          description: 所属品牌ID
          in: query
          type: integer
        - name: couSeriesId
          description: 所属系列ID
          in: query
          type: integer
        - name: couModelId
          description: 所属型号ID
          in: query
          type: integer
        - name: couWaresName
          description: 商品名称
          in: query
          type: string
        - name: status
          description: 状态 0.禁用 1.正常
          in: query
          type: integer
      responses:
        '200':
          description: 返回商品列表
          schema:
            $ref: '#/definitions/CouWaresEntities'
  /cou/wares/save:
    post:
      tags:
        - 品牌管理-商品模块
      summary: 新增商品信息
      produces:
        - application/json
      parameters:
        - name: wares
          description: 商品对象
          in: body
          schema:
            $ref: '#/definitions/CouWaresSaveVo'
          required: true
      responses:
        '200':
          description: 返回信息
          schema:
            $ref: '#/definitions/R'
  /cou/wares/info/{couWaresId}:
    get:
      tags:
        - 品牌管理-商品模块
      summary: 获取商品信息
      produces:
        - application/json
      parameters:
        - name: couWaresId
          description: 商品ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回商品对象信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              config:
                description: 返回信息
                $ref: '#/definitions/CouWaresInfoVo'
...
  /deal/wares/list:
    get:
      tags:
        - 客户商品-商品模块
      summary: 分页查询企业商品列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序字段
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: dealWaresTitle
          description: 出售商品标题
          in: query
          type: string
        - name: releaseAreaId
          description: 发布区域ID
          in: query
          type: integer
        - name: couBrandId
          description: 所属品牌ID
          in: query
          type: integer
        - name: couSeriesId
          description: 所属品牌系列ID
          in: query
          type: integer
        - name: couWaresId
          description: 所属商品ID
          in: query
          type: integer
        - name: couModelId
          description: 所属商品型号ID
          in: query
          type: integer
        - name: dealStoreId
          description: 所属客户企业ID
          in: query
          type: integer
        - name: proAreaId
          description: 所属区域省级ID
          in: query
          type: integer
        - name: cityAreaId
          description: 所属区域市级ID
          in: query
          type: integer
        - name: countyAreaId
          description: 所属区域县/区级ID
          in: query
          type: integer
        - name: onlineStatus
          description: 上线状态 0.驳回 1.销售审核中 2.经理审核中 3.上架 4.下架
          in: query
          type: integer
        - name: sellStatus
          description: 出售状态 0.未出售 1.已出售
          in: query
          type: integer
        - name: startTime
          description: 提交开始日期
          in: query
          type: string
          format: 'date-time'
        - name: endTime
          description: 提交结束日期
          in: query
          type: string
          format: 'date-time'
      responses:
        '200':
          description: 返回商品评价列表
          schema:
            $ref: '#/definitions/DealWaresEntities'
...
  /deal/wares/save:
    post:
      tags:
        - 客户商品-商品模块
      summary: 新增出售商品
      produces:
        - application/json
      parameters:
        - name: sell
          description: 出售商品对象
          in: body
          schema:
            $ref: '#/definitions/DealWaresSaveVo'
          required: true
      responses:
        '200':
          description: 返回信息
          schema:
            $ref: '#/definitions/R'
  /deal/wares/info/{dealWaresId}:
    get:
      tags:
        - 客户商品-商品模块
      summary: 根据出售商品ID获取出售商品详情
      produces:
        - application/json
      parameters:
        - name: dealWaresId
          description: 出售商品ID
          in: path
          type: String
          required: true
      responses:
        '200':
          description: 返回出售商品详情
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              config:
                description: 返回信息
                $ref: '#/definitions/DealWaresInfoVo'
...
  /deal/wares/installment/list:
    get:
      tags:
        - 客户商品-分期模块
      summary: 分页查询咨询分期列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序字段
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: dealWaresTitle
          description: 出售商品标题
          in: query
          type: string
        - name: contactName
          description: 联系人名称
          in: query
          type: string
        - name: contactPhone
          description: 联系电话
          in: query
          type: string
        - name: dealUserName
          description: 所属客户名称
          in: query
          type: string
        - name: dealUserId
          description: 所属客户ID
          in: query
          type: integer
        - name: followStatus
          description: 跟进状态 0.作废 1.待处理 2.已完成
          in: query
          type: integer
        - name: startTime
          description: 开始日期
          in: query
          type: string
          format: 'date-time'
        - name: endTime
          description: 结束日期
          in: query
          type: string
          format: 'date-time'
      responses:
        '200':
          description: 返回咨询分期列表
          schema:
            $ref: '#/definitions/DealWaresInstallmentEntities'
  /deal/wares/installment/save:
    post:
      tags:
        - 客户商品-分期模块
      summary: 新增咨询分期
      produces:
        - application/json
      parameters:
        - name: sell
          description: 咨询分期客户对象
          in: body
          schema:
            $ref: '#/definitions/DealWaresInstallmentSaveVo'
          required: true
      responses:
        '200':
          description: 返回信息
          schema:
            $ref: '#/definitions/R'
  /deal/wares/installment/info/{installmentId}:
    get:
      tags:
        - 客户商品-分期模块
      summary: 根据咨询分期ID获取咨询分期详情
      produces:
        - application/json
      parameters:
        - name: installmentId
          description: 咨询分期ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回出售商品详情
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              config:
                description: 返回信息
                $ref: '#/definitions/DealWaresInstallmentInfoVo'
  /deal/wares/installment/waste:
    post:
      tags:
        - 客户商品-分期模块
      summary: 新增修改咨询分期单跟进状态为作废
      produces:
        - application/json
      parameters:
        - name: finance
          description: 咨询分期跟进对象
          in: body
          schema:
            $ref: '#/definitions/DealInstallmentFollowInvokingVo'
          required: true
      responses:
        '200':
          description: 返回信息
          schema:
            $ref: '#/definitions/R'
  /deal/wares/installment/success:
    post:
      tags:
        - 客户商品-分期模块
      summary: 新增修改咨询分期单跟进状态为已跟进
      produces:
        - application/json
      parameters:
        - name: finance
          description: 咨询分期跟进对象
          in: body
          schema:
            $ref: '#/definitions/DealInstallmentFollowInvokingVo'
          required: true
      responses:
        '200':
          description: 返回信息
          schema:
            $ref: '#/definitions/R'

#定义数据模型
definitions:
  R:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      msg:
        description: 失败原因
        type: string
...
  SysDeptEntities:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/SysDeptListVo'
  SysDeptSaveVo:
    type: object
    properties:
      deptName:
        description: 部门名称
        type: string
      parentId:
        description: 上级部门ID
        type: integer
        format: int64
  SysDeptInfoVo:
    type: object
    properties:
      deptId:
        description: 部门ID
        type: integer
        format: int64
      deptName:
        description: 部门名称
        type: string
      parentId:
        description: 上级部门ID
        type: integer
        format: int64
  SysDeptUpdateVo:
    type: object
    properties:
      deptId:
        description: 部门ID
        type: integer
        format: int64
      deptName:
        description: 部门名称
        type: string
      parentId:
        description: 上级部门ID
        type: integer
        format: int64
...
  CouWaresEntities:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/CouWaresListVo'
  CouWaresSaveVo:
    type: object
    properties:
      couWaresName:
        description: 商品名称
        type: string
      couBrandId:
        description: 所属品牌ID
        type: integer
        format: int64
      couSeriesId:
        description: 所属系列ID
        type: integer
        format: int64
      couModelId:
        description: 所属型号ID
        type: integer
        format: int64
      couWaresPrice:
        description: 厂商指导价
        type: string
      marketYear:
        description: 年款
        type: string
      marketTime:
        description: 上市时间
        type: string
        format: date
      disMent:
        description: 排量
        type: string
      varBox:
        description: 品牌logo路径
        type: string
      drive:
        description: 品牌首字母
        type: string
      consume:
        description: 状态 0.禁用 1.正常
        type: string
  CouWaresInfoVo:
    type: object
    properties:
      couWaresId:
        description: 商品ID
        type: integer
        format: int64
      couWaresName:
        description: 商品名称
        type: string
      couBrandId:
        description: 所属品牌ID
        type: integer
        format: int64
      couBrandName:
        description: 所属品牌名称
        type: string
      couSeriesId:
        description: 所属系列ID
        type: integer
        format: int64
      couSeriesName:
        description: 所属系列名称
        type: string
      couModelId:
        description: 所属型号ID
        type: integer
        format: int64
      couModelName:
        description: 所属型号名称
        type: string
      couWaresPrice:
        description: 厂商指导价
        type: string
      marketYear:
        description: 年款
        type: string
      marketTime:
        description: 上市时间
        type: string
        format: date
      disMent:
        description: 排量
        type: string
      varBox:
        description: 变速箱
        type: string
      drive:
        description: 驱动方式
        type: string
      consume:
        description: 油耗量
        type: string
...
  DealWaresEntities:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/DealWaresListVo'
  DealWaresSaveVo:
    type: object
    properties:
      dealWaresTitle:
        description: 出售商品标题
        type: string
      releaseAreaId:
        description: 发布区域ID
        type: integer
        format: int64
      contactPhone:
        description: 联系电话
        type: string
      contactName:
        description: 联系人名称
        type: string
      sex:
        description: 性别 0.先生 1.小姐
        type: integer
        format: int64
      waresFrameCode:
        description: 商品框架号
        type: string
      couBrandId:
        description: 所属品牌ID
        type: integer
        format: int64
      couSeriesId:
        description: 所属品牌系列ID
        type: integer
        format: int64
      couWaresId:
        description: 所属商品ID
        type: integer
        format: int64
      couModelId:
        description: 所属商品型号ID
        type: integer
        format: int64
      tradePrice:
        description: 批发价
        type: string
      retailPrice:
        description: 零售价
        type: string
      registerTime:
        description: 上牌时间
        type: string
        format: date-time
      distance:
        description: 行驶里程
        type: integer
        format: int64
      licenseId:
        description: 牌照ID
        type: integer
        format: int64
      proAreaId:
        description: 省级区域ID
        type: integer
        format: int64
      cityAreaId:
        description: 市级区域ID
        type: integer
        format: int64
      countyAreaId:
        description: 县/区级区域ID
        type: integer
        format: int64
      addr:
        description: 详细地址
        type: string
      waresRemark:
        description: 商品描述
        type: string
      transferNum:
        description: 过户次数
        type: integer
        format: int64
      isTransfer:
        description: 是否含过户费 0.无 1.是
        type: integer
        format: int64
      isMortgage:
        description: 是否有抵押 0.无 1.是
        type: integer
        format: int64
      isMaintain:
        description: 是否有定期4s保养 0.无 1.是
        type: integer
        format: int64
      coverImage:
        description: 商品封面图
        type: object
        properties:
          image:
            description: 图片路径
            type: string
      driveImage:
        description: 行驶证图
        type: object
        properties:
          image:
            description: 图片路径
            type: string
      waresImages:
        description: 商品评估图集合
        type: array
        items:
          image:
            description: 图片路径
            type: string
      dealStoreId:
        description: 所属企业客户ID
        type: integer
        format: int64
  DealWaresInfoVo:
    type: object
    properties:
      dealWaresId:
        description: 出售商品ID
        type: string
      dealWaresTitle:
        description: 出售商品标题
        type: string
      dealWaresNo:
        description: 出售商品编号
        type: string
      releaseAreaId:
        description: 发布区域ID
        type: integer
        format: int64
      releaseAreaName:
        description: 发布区域名称
        type: string
      contactPhone:
        description: 联系电话
        type: string
      contactName:
        description: 联系人名称
        type: string
      sex:
        description: 性别 0.先生 1.小姐
        type: integer
        format: int64
      waresFrameCode:
        description: 商品框架号
        type: string
      couBrandId:
        description: 所属品牌ID
        type: integer
        format: int64
      couBrandName:
        description: 所属品牌名称
        type: string
      couSeriesId:
        description: 所属品牌系列ID
        type: integer
        format: int64
      couSeriesName:
        description: 所属品牌系列名称
        type: string
      couWaresId:
        description: 所属商品ID
        type: integer
        format: int64
      couWaresName:
        description: 所属商品名称
        type: string
      couModelId:
        description: 所属商品型号ID
        type: integer
        format: int64
      couModelName:
        description: 所属商品型号名称
        type: string
      tradePrice:
        description: 批发价
        type: string
      retailPrice:
        description: 零售价
        type: string
      registerTime:
        description: 上牌时间
        type: string
        format: date-time
      distance:
        description: 行驶里程
        type: integer
        format: int64
      licenseId:
        description: 牌照ID
        type: integer
        format: int64
      licenseCode:
        description: 牌照
        type: string
      proAreaId:
        description: 省级区域ID
        type: integer
        format: int64
      cityAreaId:
        description: 市级区域ID
        type: integer
        format: int64
      countyAreaId:
        description: 县/区级区域ID
        type: integer
        format: int64
      addr:
        description: 详细地址
        type: string
      waresRemark:
        description: 商品描述
        type: string
      transferNum:
        description: 过户次数
        type: integer
        format: int64
      isTransfer:
        description: 是否含过户费 0.无 1.是
        type: integer
        format: int64
      isMortgage:
        description: 是否有抵押 0.无 1.是
        type: integer
        format: int64
      isMaintain:
        description: 是否有定期4s保养 0.无 1.是
        type: integer
        format: int64
      sellStatus:
        description: 出售状态 0.未出售 1.已出售
        type: integer
        format: int64
      onlineStatus:
        description: 上线状态 0.驳回 1.销售审核中 2.经理审核中 3.上架 4.下架
        type: integer
        format: int64
      coverImage:
        type: object
        properties:
          imageId:
            description: 图片ID
            type: integer
            format: int64
          image:
            description: 图片路径
            type: string
          imageType:
            description: 图片类型 0.商品图 1.行驶证图
            type: integer
            format: int64
          isCover:
            description: 是否封面图 0.否 1.是
            type: integer
            format: int64
          dealWaresId:
            description: 所属企业商品ID
            type: integer
            format: int64
      driveImage:
        type: object
        properties:
          imageId:
            description: 图片ID
            type: integer
            format: int64
          image:
            description: 图片路径
            type: string
          imageType:
            description: 图片类型 0.商品图 1.行驶证图
            type: integer
            format: int64
          isCover:
            description: 是否封面图 0.否 1.是
            type: integer
            format: int64
          dealWaresId:
            description: 所属企业商品ID
            type: integer
            format: int64
      waresImages:
        type: array
        items:
          imageId:
            description: 图片ID
            type: integer
            format: int64
          image:
            description: 图片路径
            type: string
          imageType:
            description: 图片类型 0.商品图 1.行驶证图
            type: integer
            format: int64
          isCover:
            description: 是否封面图 0.否 1.是
            type: integer
            format: int64
          dealWaresId:
            description: 所属企业商品ID
            type: integer
            format: int64
      dealStoreId:
        description: 所属企业客户ID
        type: integer
        format: int64
      dealUserName:
        description: 所属企业客户名称
        type: string
      submitTime:
        description: 提交时间
        type: string
        format: date-time
...
  DealWaresInstallmentEntities:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/DealWaresInstallmentListVo'
  DealWaresInstallmentSaveVo:
    type: object
    properties:
      dealUserId:
        description: 所属客户ID
        type: integer
        format: int64
      dealWaresId:
        description: 咨询商品ID
        type: string
      contactPhone:
        description: 联系电话
        type: string
      contactName:
        description: 联系人名称
        type: string
      sex:
        description: 性别 0.先生 1.小姐
        type: integer
        format: int64
...
  DealInstallmentFollowInvokingVo:
    type: object
    properties:
      installmentId:
        description: 咨询分期ID
        type: integer
        format: int64
      followRemake:
        description: 跟进备注
        type: string
...